requirements.txt
flask>=3.0
flask-sqlalchemy>=3.1
argon2-cffi>=23.1
cryptography>=42.0
pyotp>=2.9
qrcode[pil]>=7.4
authlib>=1.3
python-dotenv>=1.0
requests>=2.31
